+++ src/tiledb/cloud/utilities/consolidate.py
@@ -101,7 +101,9 @@ def consolidate_and_vacuum(
     logger = get_logger()
 
     with tiledb.scope_ctx(config):
-        if vacuum_fragments:
+        # TODO: remove when remote fragment consolidation is supported
+        is_remote = array_uri.startswith("tiledb://")
+        if not is_remote and vacuum_fragments:
             logger.info("Vacuuming fragments")
             tiledb.vacuum(
                 array_uri,