+++ tests/test_file.py
@@ -277,10 +277,13 @@ class UploadFolderTest(unittest.TestCase):
         )
 
 
+@unittest.skip("Skip until fixed VFS access")
 class TestFileIngestion(unittest.TestCase):
     @classmethod
     def setUpClass(cls) -> None:
-        """Setup test files, group and destinations once before the file tests start."""
+        """
+        Setup test files, group and destinations once before the file tests start.
+        """
         cls.input_file_location = (
             "s3://tiledb-unittest/groups/file_ingestion_test_files"
         )
@@ -313,6 +316,13 @@ class TestFileIngestion(unittest.TestCase):
     @classmethod
     def tearDownClass(cls) -> None:
         """Cleanup after the tests have run"""
+        cls.cleanup_these_uris += [
+            f"tiledb://{cls.namespace}/{gifn}" for gifn in cls.group_input_file_names
+        ]
+        cls.cleanup_these_uris += [
+            f"tiledb://{cls.namespace}/{ifn}" for ifn in cls.input_file_names
+        ]
+
         _cleanup_residual_test_arrays(array_uris=cls.cleanup_these_uris)
         groups.delete(cls.group_uri, recursive=True)
         return super().tearDownClass()
@@ -358,6 +368,11 @@ class TestFileIngestion(unittest.TestCase):
 
         # Clean up
         _cleanup_residual_test_arrays(array_uris=ingested_array_uris)
+        _cleanup_residual_test_arrays(
+            array_uris=[
+                f"tiledb://{self.namespace}/{iau}" for iau in ingested_array_uris
+            ]
+        )
 
     def test_add_array_to_group_udf_raises_bad_namespace_error(self):
         with self.assertRaises(tiledb.TileDBError):
@@ -378,10 +393,13 @@ class TestFileIngestion(unittest.TestCase):
             )
 
 
+@unittest.skip("Skip until fixed VFS access")
 class TestFileIndexing(unittest.TestCase):
     @classmethod
     def setUpClass(cls) -> None:
-        """Setup test files, group and destinations once before the file tests start."""
+        """
+        Setup test files, group and destinations once before the file tests start.
+        """
         cls.input_file_location = "s3://tiledb-unittest/groups/file_indexing_test_files"
         # Files with name "input_file_<n[0, 4]>.pdf" have already been placed
         # in the "cls.input_file_location"