+++ CHANGES.md
@@ -10,6 +10,7 @@ New features:
 
 Bug fixes:
 
+- Enable ingestion of multi-band geospatial raster data (gh-609).
 - Ensure that boolean recursive parameters in tiledb.cloud.asset and
   tiledb.cloud.groups are converted to "true" or "false" before server requests
   are made (gh-605).